% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atrous_dwt.R
\name{atrous_dwt}
\alias{atrous_dwt}
\title{A Trous Discrete Wavelet Transform}
\usage{
atrous_dwt(X, wavelet, decomp_level)
}
\arguments{
\item{X}{Time series (N x 1)}

\item{wavelet}{Scaling filter name (string)}

\item{decomp_level}{Decomposition level (1 < integer < N/2)}
}
\value{
W - wavelet and scaling coefficients (N x J+1) (wavelet coefs in first J columns)
}
\description{
This function calculates the wavelet and scaling coefficients of the a trous (AT)
version of the Discrete Wavelet Transform (DWT).
}
\examples{
\dontrun{
N = 1000 #  number of time series points
J = 4 # decomposition level
wavelet = 'coif1' # wavelet filter
X = matrix(rnorm(N),N,1)
W = atrous_dwt(X,wavelet,J)
Xr = as.matrix(rowSums(W)) # reconstruct time series
mse_r = mean( (X - Xr)^2) # confirm additive reconstruction
plot.ts(W) # plot wavelet and scaling coefficients
}
}
\references{
Benaouda, D., F. Murtagh, J. L. Starck, and O. Renaud (2006),
Wavelet-based nonlinear multiscale decomposition model for
electricity load forecasting, Neurocomputing,
doi:10.1016/j.neucom.2006.04.005.

Maheswaran, R., and R. Khosa (2012), Comparative study of different
wavelets for hydrologic forecasting, Comput. Geosci.,
doi:10.1016/j.cageo.2011.12.015.
}
