% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circular_convolution}
\alias{circular_convolution}
\title{Circular Convolution}
\usage{
circular_convolution(a, b)
}
\arguments{
\item{a}{Filter (typically, the wavelet or scaling filter)}

\item{b}{A vector}
}
\value{
Result of circular convolution, length is the same as \code{length(b)}
}
\description{
Perform a circular convolution between a filter and a time series.
}
\details{
The convolution theorem tells us that the discrete Fourier transform
can be leveraged to compute a circular convolution, which is what this function does.
}
\seealso{
\code{\link[stats]{fft}}
}
