% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{equivalent_filters}
\alias{equivalent_filters}
\title{Equivalent Wavelet and Scaling Filters}
\usage{
equivalent_filters(filter, J, modwt = FALSE)
}
\arguments{
\item{filter}{Filter name (string)}

\item{J}{Decomposition level (integer)}

\item{modwt}{Do you want the MODWT version?, Default: FALSE}
}
\value{
A list containing the equivalent wavelet and scaling filter and their lengths.
}
\description{
Equivalent Wavelet and Scaling Filters
}
\details{
Equivalent wavelet and scaling filters are useful for obtaining
MODWT wavelet and scaling coefficients for a given filter name and
decomposition level.
}
