% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_dwt.R
\name{mo_dwt}
\alias{mo_dwt}
\title{Maximal Overlap Discrete Wavelet Transform (MODWT)}
\usage{
mo_dwt(X, wavelet, decomp_level)
}
\arguments{
\item{X}{Time series (N x 1)}

\item{wavelet}{Scaling filter name (string)}

\item{decomp_level}{Decomposition level (1 < integer < N/2)}
}
\value{
Wavelet and scaling coefficients (N x J+1) (wavelet coefs in first J columns)
}
\description{
This function calculates the wavelet and scaling coefficients of the
MODWT.
}
\examples{
\dontrun{

source('H:/Rfolder/Functions and Scripts/scaling_filter.R')
source('H:/Rfolder/Functions and Scripts/scaling_coefs.R')
source('H:/Rfolder/Functions and Scripts/wavelet_coefs.R')
source('H:/Rfolder/Functions and Scripts/wavelet_filter.R')
source('H:/Rfolder/Functions and Scripts/n_boundary_coefs.R')
source('H:/Rfolder/Functions and Scripts/mo_dwt.R')
N = 1000 #  number of time series points
J = 4 # decomposition level
wavelet = 'coif1' # wavelet filter
nbc = n_boundary_coefs(wavelet,J) # number of boundary-effected coefficients at decomp_level J
X = matrix(rnorm(N),N,1)
W = mo_dwt(X,wavelet,J)
}
}
\references{
M. Basta (2014), Additive Decomposition and Boundary Conditions in Wavelet-Based
Forecasting Approaches, Acta Oeconomica Pragensia, 2, pp. 48-70.
}
