% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{modwt_wavelet_coefs}
\alias{modwt_wavelet_coefs}
\alias{modwt_scaling_coefs}
\title{MODWT Wavelet Coefficients}
\usage{
modwt_wavelet_coefs(X, filter, j)

modwt_scaling_coefs(X, filter, j)
}
\arguments{
\item{X}{A vector}

\item{filter}{Filter name (string)}

\item{j}{Decomposition level (integer)}
}
\value{
MODWT wavelet coefficients for specified decomposition level (vector)

MODWT scaling coefficients for specified decomposition level (vector)
}
\description{
Compute MODWT wavelet coefficients for a specified filter and
decomposition level. This function only gives the MODWT wavelet coefficients
for the user specified level. To get all MODWT wavelet coefficients up to and
including the user specified decomposition level (as well as the MODWT scaling
coefficient corresponding to the user specified decomposition level) all in one nice
matrix, please use \code{mo_dwt()}.

Compute MODWT scaling coefficients for a specified filter and
decomposition level. This function only gives the MODWT scaling coefficients
for the user specified level.
}
\details{
This function implements the first equation in equation 169a (page 169)
of Wavelet Methods for Time Series Analysis by Percival and Walden.

This function implements the second equation in equation 169a (page 169)
of Wavelet Methods for Time Series Analysis by Percival and Walden.
}
