% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_boundary_coefs.R
\name{n_boundary_coefs}
\alias{n_boundary_coefs}
\title{Number of Boundary Coefficients}
\usage{
n_boundary_coefs(wavelet, decomp_level)
}
\arguments{
\item{wavelet}{scaling filter name [string]}

\item{decomp_level}{decomposition level [integer]}
}
\value{
number of boundary coefficients [integer]
}
\description{
This function calculates the number of boundary coefficients for a
particular wavelet/scaling filter and decomposition level.
}
\examples{
\dontrun{
J = 4 # decomposition level
wavelet = 'b3spline' # wavelet filter
nbc = n_boundary_coefs(wavelet,J) # number of boundary-effected coefficients at decomp_level J
}
}
\references{
M. Basta (2014),Additive Decomposition and Boundary Conditions in Wavelet-Based
Forecasting Approaches, Acta Oeconomica Pragensia, 2, pp. 48-70.
}
