% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{periodize_filter}
\alias{periodize_filter}
\title{Periodize a filter}
\usage{
periodize_filter(filter_vector, periodize_to)
}
\arguments{
\item{filter_vector}{The filter vector that is outputted from r_wavelet_filter or r_scaling_filter}

\item{periodize_to}{The length of the time series that is to be decomposed}
}
\value{
Periodized filter
}
\description{
Periodize a filter
}
\details{
Output of the function is a zero padded version of the original filter, padded at the end.
}
