% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{r_scaling_filter}
\alias{r_scaling_filter}
\title{Scaling Filter}
\usage{
r_scaling_filter(filter, modwt = FALSE)
}
\arguments{
\item{filter}{Name of the scaling filter desired [string]}

\item{modwt}{Should the elements of the returned filter be scaled by \code{1/sqrt(2)}?, Default: \code{FALSE}}
}
\value{
A scaling filter vector
}
\description{
This function returns the user specified scaling filter.
}
\details{
See equation 75a (page 75) of Wavelet Methods for Time Series Analysis
by Percival and Walden to see how to convert a wavelet filter into a scaling filter.

The argument \code{filter} can take one of the following values:

c('bl7', 'bl9', 'bl10',
'beyl',
'coif1', 'coif2', 'coif3', 'coif4', 'coif5',
'db1', 'db2', 'db3', 'db4', 'db5', 'db6', 'db7', 'db8', 'db9', 'db10', 'db11', 'db12',
'db13', 'db14', 'db15', 'db16', 'db17', 'db18', 'db19', 'db20', 'db21', 'db22', 'db23',
'db24', 'db25', 'db26', 'db27', 'db28', 'db29', 'db30', 'db31', 'db32', 'db33',
'db34', 'db35', 'db36', 'db37', 'db38', 'db39', 'db40', 'db41', 'db42', 'db43', 'db44', 'db45',
'fk4', 'fk6', 'fk8', 'fk14', 'fk18', 'fk22',
'han2_3', 'han3_3', 'han4_5', 'han5_5',
'dmey',
'mb4_2', 'mb8_2', 'mb8_3', 'mb8_4', 'mb10_3', 'mb12_3', 'mb14_3', 'mb16_3', 'mb18_3', 'mb24_3', 'mb32_3',
'sym2', 'sym3', 'sym4', 'sym5', 'sym6', 'sym7', 'sym8', 'sym9', 'sym10', 'sym11', 'sym12', 'sym13', 'sym14',
'sym15', 'sym16', 'sym17', 'sym18', 'sym19', 'sym20', 'sym21', 'sym22', 'sym23', 'sym24', 'sym25', 'sym26', 'sym27',
'sym28', 'sym29', 'sym30', 'sym31', 'sym32', 'sym33', 'sym34', 'sym35', 'sym36', 'sym37', 'sym38', 'sym39', 'sym40',
'sym41', 'sym42', 'sym43', 'sym44', 'sym45',
'vaid',
'la8', 'la10', 'la12', 'la14', 'la16', 'la18', 'la20')
}
\examples{
\dontrun{
if(interactive()){
 r_scaling_filter("db1")
 }
}
}
