% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaling_coefs}
\alias{scaling_coefs}
\title{Compute Scaling Coefficients}
\usage{
scaling_coefs(X, wavelet, j)
}
\arguments{
\item{X}{An (N x 1) matrix or a vector}

\item{wavelet}{A character string indicating the scaling filter}

\item{j}{The decomposition level [integer]}
}
\value{
An (N x 1) matrix scaling coefficients
}
\description{
Compute the scaling coefficients.
}
\references{
Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge
University Press.
}
