% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_check.R
\name{shape_check}
\alias{shape_check}
\title{Check Matrix X is (N x 1)}
\usage{
shape_check(X)
}
\arguments{
\item{X}{Object to check and (if possible) coerce to a single column matrix}
}
\value{
A single column matrix
}
\description{
\code{shape_check} checks whether \code{X} is a matrix representing
a column vector (i.e., a matrix with 1 column). If not, \code{shape_check} attempts
to coerce the user provided \code{X} to a matrix with 1 column. If this cannot be done,
an error is raised.
}
\details{
This is a utility function written to check the input \code{X} for the
functions \code{atrous_dwt} and \code{mo_dwt}.
}
