% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wavelet_coefs}
\alias{wavelet_coefs}
\title{Compute Wavelet Coefficients}
\usage{
wavelet_coefs(X, wavelet, j)
}
\arguments{
\item{X}{A numeric vector}

\item{wavelet}{A character string indicating the scaling filter desired}

\item{j}{The decomposition level [integer]}
}
\value{
Matrix of wavelet coefficients
}
\description{
Compute Wavelet Coefficients
}
\references{
Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge
University Press.
}
