% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wavelet_filter}
\alias{wavelet_filter}
\title{Compute the wavelet filter}
\usage{
wavelet_filter(wavelet)
}
\arguments{
\item{wavelet}{A character string indicating the wavelet filter desired}
}
\value{
Wavelet filter vector
}
\description{
Compute the wavelet filter
}
\references{
Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge
University Press.
}
