% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{n_boundary_coefs}
\alias{n_boundary_coefs}
\title{Number of Boundary Coefficients}
\usage{
n_boundary_coefs(filter, J)
}
\arguments{
\item{filter}{Filter name (see Details below) [string]}

\item{J}{Decomposition level [integer]}
}
\value{
Number of boundary coefficients [integer]
}
\description{
This function calculates the number of boundary coefficients for a
particular wavelet/scaling filter and decomposition level.
}
\details{
The argument \code{filter} can take one of the following values:

c('bl7', 'bl9', 'bl10',
'beyl',
'coif1', 'coif2', 'coif3', 'coif4', 'coif5',
'db1', 'db2', 'db3', 'db4', 'db5', 'db6', 'db7', 'db8', 'db9', 'db10', 'db11', 'db12',
'db13', 'db14', 'db15', 'db16', 'db17', 'db18', 'db19', 'db20', 'db21', 'db22', 'db23',
'db24', 'db25', 'db26', 'db27', 'db28', 'db29', 'db30', 'db31', 'db32', 'db33',
'db34', 'db35', 'db36', 'db37', 'db38', 'db39', 'db40', 'db41', 'db42', 'db43', 'db44', 'db45',
'fk4', 'fk6', 'fk8', 'fk14', 'fk18', 'fk22',
'han2_3', 'han3_3', 'han4_5', 'han5_5',
'dmey',
'mb4_2', 'mb8_2', 'mb8_3', 'mb8_4', 'mb10_3', 'mb12_3', 'mb14_3', 'mb16_3', 'mb18_3', 'mb24_3', 'mb32_3',
'sym2', 'sym3', 'sym4', 'sym5', 'sym6', 'sym7', 'sym8', 'sym9', 'sym10', 'sym11', 'sym12', 'sym13', 'sym14',
'sym15', 'sym16', 'sym17', 'sym18', 'sym19', 'sym20', 'sym21', 'sym22', 'sym23', 'sym24', 'sym25', 'sym26', 'sym27',
'sym28', 'sym29', 'sym30', 'sym31', 'sym32', 'sym33', 'sym34', 'sym35', 'sym36', 'sym37', 'sym38', 'sym39', 'sym40',
'sym41', 'sym42', 'sym43', 'sym44', 'sym45',
'vaid',
'la8', 'la10', 'la12', 'la14', 'la16', 'la18', 'la20')
}
\examples{
J <- 4 # decomposition level
nbc <- n_boundary_coefs('db1', J) # number of boundary-effected coefficients at decomp_level J
}
\references{
M. Basta (2014),Additive Decomposition and Boundary Conditions in Wavelet-Based
Forecasting Approaches, Acta Oeconomica Pragensia, 2, pp. 48-70.

Quilty, J., & Adamowski, J. (2018). Addressing the incorrect usage of wavelet-based
hydrological and water resources forecasting models for real-world applications with best
practices and a new forecasting framework. Journal of Hydrology, 563, 336–353.
https://doi.org/10.1016/j.jhydrol.2018.05.003

Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge
University Press.
}
