% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaling_coefs}
\alias{scaling_coefs}
\title{Compute Scaling Coefficients}
\usage{
scaling_coefs(X, wavelet, j)
}
\arguments{
\item{X}{A numeric vector}

\item{wavelet}{A character string indicating the scaling filter desired}

\item{j}{The decomposition level [integer]}
}
\value{
Matrix of scaling coefficients
}
\description{
Compute Scaling Coefficients
}
\references{
Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge
University Press.
}
