% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaling_coefs}
\alias{scaling_coefs}
\title{Computes scaling coefficients}
\usage{
scaling_coefs(X, wavelet, j)
}
\arguments{
\item{X}{A vector}

\item{wavelet}{A character string indicating the wavelet filter desired}

\item{j}{The decomposition level}
}
\value{
Matrix of scaling coefficients
}
\description{
Computes scaling coefficients
}
