% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaling_filter}
\alias{scaling_filter}
\title{Computes the scaling filter}
\usage{
scaling_filter(wavelet)
}
\arguments{
\item{wavelet}{A character string indicating the wavelet filter desired}
}
\value{
Scaling filter vector
}
\description{
Computes the scaling filter
}
