% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaling_filter}
\alias{scaling_filter}
\title{Maximal Overlap Discrete Wavelet Transform Scaling Filter}
\usage{
scaling_filter(filter)
}
\arguments{
\item{filter}{A character string indicating the scaling filter desired}
}
\value{
Scaling filter vector (a numeric vector)
}
\description{
Compute the MODWT scaling filter, which is just the regular wavelet
filter, divided by \code{sqrt(2)}. This function is used for the pyramid algorithm
used in \code{modwt}.
}
\references{
Percival, D. B. and A. T. Walden (2000) Wavelet Methods for Time Series Analysis, Cambridge
University Press.

Wasilewski, F. (2008). Wavelet browser by pywavelets. Wavelet Properties Browser.
Retrieved November 17, 2022, from http://wavelets.pybytes.com/

Gregory R. Lee, Ralf Gommers, Filip Wasilewski, Kai Wohlfahrt, Aaron O’Leary (2019).
PyWavelets: A Python package for wavelet analysis. Journal of Open Source Software,
4(36), 1237, https://doi.org/10.21105/joss.01237.

Olhede, S., & Walden, A. T. (2004). The Hilbert spectrum via wavelet projections.
Proceedings of the Royal Society of London. Series A: Mathematical, Physical and
Engineering Sciences, 460(2044), 955–975. https://doi.org/10.1098/rspa.2003.1199

Maheswaran, R., & Khosa, R. (2012). Comparative study of different wavelets for
hydrologic forecasting. Computers & Geosciences, 46, 284–295.
https://doi.org/10.1016/j.cageo.2011.12.015

Daubechies, Ingrid. Ten Lectures on Wavelets. Society for Industrial and Applied Mathematics, 1992.

Morris, Joel M, and Ravindra Peravali. “Minimum-Bandwidth Discrete-Time Wavelets.” Signal Processing 76,
no. 2 (July 1999): 181–93. https://doi.org/10.1016/S0165-1684(99)00007-9.

Doroslovački, M.L. “On the Least Asymmetric Wavelets.” IEEE Transactions on Signal Processing 46,
no. 4 (April 1998): 1125–30. https://doi.org/10.1109/78.668562.

Han, Bin. “Wavelet Filter Banks.” In Framelets and Wavelets: Algorithms, Analysis, and Applications,
92–98. Applied and Numerical Harmonic Analysis. Cham, Switzerland: Birkhäuser, 2017.
https://doi.org/10.1007/978-3-319-68530-4_2.

Vaidyanathan, P. P. & Hoang, P.-Q. Lattice structures for optimal design and robust implementation
of two-channel perfect-reconstruction QMF banks. IEEE Trans. Acoust. 36, 81–94 (1988).

Wickerhauser, M. V. Adapted Wavelet Analysis from Theory to Software.
Adapted Wavelet Analysis from Theory to Software (A.K. Peters, 1994).
}
\keyword{internal}
