% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wavelet_coefs}
\alias{wavelet_coefs}
\title{Computes wavelet coefficients}
\usage{
wavelet_coefs(X, wavelet, j)
}
\arguments{
\item{X}{A vector}

\item{wavelet}{A character string indicating the wavelet filter desired}

\item{j}{The decomposition level}
}
\value{
Matrix of wavelet coefficients
}
\description{
Computes wavelet coefficients
}
