% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wavelet_filter}
\alias{wavelet_filter}
\title{Computes the wavelet filter}
\usage{
wavelet_filter(wavelet)
}
\arguments{
\item{wavelet}{A character string indicating the wavelet filter desired}
}
\value{
Wavelet filter vector
}
\description{
Computes the wavelet filter
}
